requirements.txt
alembic==1.18.4
blinker==1.9.0
charset-normalizer==3.4.4
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
defusedxml==0.7.1
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.55.0
fpdf2==2.8.5
greenlet==3.3.1
gunicorn==25.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.6
livereload==2.7.1
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.9.3
numpy==2.1.3
packaging==24.2
pillow==11.0.0
pyparsing==3.1.4
python-dateutil==2.9.0
python-dotenv==1.2.1
reportlab==4.4.10
setuptools==70.3.0
six==1.16.0
SQLAlchemy==2.0.46
tornado==6.5.4
typing_extensions==4.15.0
Werkzeug==3.0.6
WTForms==3.2.1
